requirements.txt
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.4
django-cors-headers==2.1.0
django-extensions==1.9.0
djangorestframework==3.6.4
gunicorn==19.7.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycurl==7.43.0
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
